Change i128 to smaller types where it's e.g. days or bigger.
2**63 days is 2e16 years, we don't need more than that.

Maybe it matters for seconds? We need to set up some general
expectation for what range of years is reasonable and try
to choose data types so that we stick inside of that range.

+++ src/iso8601/date_time_builder.rs
@@ -10,7 +10,7 @@ use crate::zoneinfo::get_leap_seconds;
 pub struct DateTimeBuilder {
     chronology: Option<Chronology>,
     precision: Option<Precision>,
-    year: Option<i128>,
+    year: Option<u16>,
     month: Option<u8>,
     day: Option<u8>,
     hour: Option<u8>,
@@ -33,7 +33,7 @@ impl DateTimeBuilder {
         self
     }
 
-    pub fn year(&mut self, year: i128) -> &mut Self {
+    pub fn year(&mut self, year: u16) -> &mut Self {
         self.year = Some(year);
         self.precision = opt_max(self.precision, Precision::Years);
         self
